+++ packages/theme/etc/theme.api.md
@@ -9,6 +9,12 @@ import { IRawStyle } from '@uifabric/merge-styles';
 import { IStyleFunctionOrObject } from '@uifabric/utilities';
 
 // @public
+export const AnimationStyles: IAnimationStyles;
+
+// @public
+export const AnimationVariables: IAnimationVariables;
+
+// @public
 export type ColorTokens = Partial<{
     background: string;
     contentColor: string;
@@ -153,6 +159,102 @@ export namespace FontWeights {
     semibold: IFontWeight;
     const // (undocumented)
     bold: IFontWeight;
+}
+
+// @public
+export interface IAnimationStyles {
+    // (undocumented)
+    fadeIn100: IRawStyle;
+    // (undocumented)
+    fadeIn200: IRawStyle;
+    // (undocumented)
+    fadeIn400: IRawStyle;
+    // (undocumented)
+    fadeIn500: IRawStyle;
+    // (undocumented)
+    fadeOut100: IRawStyle;
+    // (undocumented)
+    fadeOut200: IRawStyle;
+    // (undocumented)
+    fadeOut400: IRawStyle;
+    // (undocumented)
+    fadeOut500: IRawStyle;
+    // (undocumented)
+    rotate90deg: IRawStyle;
+    // (undocumented)
+    rotateN90deg: IRawStyle;
+    // (undocumented)
+    scaleDownIn100: IRawStyle;
+    // (undocumented)
+    scaleDownOut98: IRawStyle;
+    // (undocumented)
+    scaleUpIn100: IRawStyle;
+    // (undocumented)
+    scaleUpOut103: IRawStyle;
+    // (undocumented)
+    slideDownIn10: IRawStyle;
+    // (undocumented)
+    slideDownIn20: IRawStyle;
+    // (undocumented)
+    slideDownOut10: IRawStyle;
+    // (undocumented)
+    slideDownOut20: IRawStyle;
+    // (undocumented)
+    slideLeftIn10: IRawStyle;
+    // (undocumented)
+    slideLeftIn20: IRawStyle;
+    // (undocumented)
+    slideLeftIn40: IRawStyle;
+    // (undocumented)
+    slideLeftIn400: IRawStyle;
+    // (undocumented)
+    slideLeftOut10: IRawStyle;
+    // (undocumented)
+    slideLeftOut20: IRawStyle;
+    // (undocumented)
+    slideLeftOut40: IRawStyle;
+    // (undocumented)
+    slideLeftOut400: IRawStyle;
+    // (undocumented)
+    slideRightIn10: IRawStyle;
+    // (undocumented)
+    slideRightIn20: IRawStyle;
+    // (undocumented)
+    slideRightIn40: IRawStyle;
+    // (undocumented)
+    slideRightIn400: IRawStyle;
+    // (undocumented)
+    slideRightOut10: IRawStyle;
+    // (undocumented)
+    slideRightOut20: IRawStyle;
+    // (undocumented)
+    slideRightOut40: IRawStyle;
+    // (undocumented)
+    slideRightOut400: IRawStyle;
+    // (undocumented)
+    slideUpIn10: IRawStyle;
+    // (undocumented)
+    slideUpIn20: IRawStyle;
+    // (undocumented)
+    slideUpOut10: IRawStyle;
+    // (undocumented)
+    slideUpOut20: IRawStyle;
+}
+
+// @public (undocumented)
+export interface IAnimationVariables {
+    // (undocumented)
+    durationValue1: string;
+    // (undocumented)
+    durationValue2: string;
+    // (undocumented)
+    durationValue3: string;
+    // (undocumented)
+    durationValue4: string;
+    // (undocumented)
+    easeFunction1: string;
+    // (undocumented)
+    easeFunction2: string;
 }
 
 // @public (undocumented)
