+++ packages/theme/src/motions/index.ts
@@ -1,1 +1,2 @@
 export * from './FluentMotions';
+export * from './AnimationStyles';
