+++ packages/theme/src/types/index.ts
@@ -4,5 +4,6 @@ export { IPalette } from './IPalette';
 export { ISemanticColors } from './ISemanticColors';
 export { ISemanticTextColors } from './ISemanticTextColors';
 export { ISpacing } from './ISpacing';
+export { IAnimationStyles, IAnimationVariables } from './IAnimationStyles';
 export { ITheme, IPartialTheme, IScheme, ISchemeNames } from './ITheme';
 export * from './Theme';
